Removing unneeded files, creating shared types file, typescript fixes
