--- old files/lizzyImporter.html
@@ -1,10 +1,0 @@
-<html>
-  <body>
-    <h1>Lizzy Import App</h1>
-    <button id="viewSavedBasketsBtn">View Baskets</button>
-    <button id="importPartListBtn">Import Parts</button>
-    <button id="clearStorageBtn">Clear Storage</button>
-    <div id="basketList"></div>
-    <script src="scripts/baskets.js" defer></script>
-  </body>
-</html>
