--- old files/manifest3.json
@@ -1,46 +1,0 @@
-{
-  "manifest_version": 3,
-  "name": "Lizzy Importer with React",
-  "version": "1.0",
-  "description": "Exports information from triumph and allows user to import part information into Lizzy",
-  "options_page": "options.html",
-  "background": {
-    "service_worker": "background.bundle.js"
-  },
-  "action": {
-    "default_popup": "popup.html",
-    "default_icon": "icon-34.png"
-  },
-  "icons": {
-    "128": "icon-128.png"
-  },
-  "content_scripts": [
-    {
-      "js": [
-        "contentScript.bundle.js"
-      ],
-      "matches": [
-        "https://www.triumphonline.net/EPC/Basket.aspx*/*"
-      ],
-      "css": [
-        "content.styles.css"
-      ]
-    }
-  ],
-  "web_accessible_resources": [
-    {
-      "resources": [
-        "content.styles.css",
-        "icon-128.png",
-        "icon-34.png"
-      ],
-      "matches": [
-        "https://www.triumphonline.net/*"
-      ]
-    }
-  ],
-  "permissions": [
-    "activeTab",
-    "storage"
-  ]
-}
