--- old files/scripts/part.js
@@ -1,10 +1,0 @@
-class Part {
-  constructor(params) {
-    const { partNumber, quantity, description } = params;
-    this.number = partNumber || null;
-    this.description = description || '';
-    this.quantity = quantity || 0;
-  }
-}
-
-export default Part;
