+++ pages/content/src/importer.ts
@@ -1,52 +1,20 @@
+import { Basket, ImportBasketMessage } from '../../../chrome-extension/public/types';
 // constant variables that may have to change in the future but leaving as default 0 for now
 const invoice_type = 0;
 const globalInvoiceID = 0;
 
-type Basket = {
-  id: string;
-  name: string;
-  partList: Part[];
-};
-
-interface Part {
-  number: string;
-  quantity: number;
-  description: string;
-}
-
-interface Message {
-  type: string;
-}
-
-interface ImportBasketMessage extends Message {
-  type: 'IMPORT_BASKET';
-  basket: Basket;
-}
-
 let itemLocatorIntervalId: NodeJS.Timeout;
 let count = 0;
 
-const importBasket = (basket: Basket, sendResponse) => {
+const importBasket = (basket: Basket) => {
   // TODO: check if the screen we expect to be up currently is displayed if not error
   // TOOD: FUTURE: if the screen isn't up try to display it
-  console.log(basket);
-
   addItemsToBasket(basket);
 };
 
 const addItemsToBasket = async (basket: Basket) => {
-  /**
-   * Functions:
-   * get_cv_itemslist
-   */
   const { partList } = basket;
-  /**
-     * {
-    "description": "Mount Boss, Screen Handle ",
-    "number": "T2319108 ",
-    "quantity": 1
-}
-     */
+
   for (const part of partList) {
     const { number, quantity } = part;
 
@@ -78,13 +46,11 @@ const addItemsToBasket = async (basket: Basket) => {
     }
 
     const href = itemSelectorLinkElement.getAttribute('href') || '';
-    const { itemId, childItemId } = /(?<ignore1>\D+)(?<itemId>\d+)(?<ignore2>\D+)(?<childItemId>\d+)/.exec(href).groups;
+    const { itemId, childItemId } =
+      /(?<ignore1>\D+)(?<itemId>\d+)(?<ignore2>\D+)(?<childItemId>\d+)/.exec(href)?.groups || {};
 
-    // TODO: check if SS
-    // display the item
-    const response = await invItemSelect('reg_itemlocate', itemId, false, childItemId);
+    const response = await invItemSelect('reg_itemlocate', itemId, false, parseInt(childItemId) ?? 0);
 
-    // I can't click on the link in popup so follow code path and figure out what I need to do to get this to work
     if (response) {
       await call_cs_invoice_items(response);
     } else {
@@ -99,8 +65,6 @@ const addItemsToBasket = async (basket: Basket) => {
     }
     partQuantityInput.value = quantity.toString();
 
-    // // save the part and quantity
-    // window.save_cv_invoice_itemlocate('frm_invoice_itemlocate', 1, 0, false, 'open');
     const addButtonElement = document.querySelector('#btn_additem') as HTMLButtonElement;
     if (!addButtonElement) {
       console.error('Add button element not found');
@@ -117,7 +81,7 @@ export const addMessageListener = () => {
       switch (message.type) {
         case 'IMPORT_BASKET':
           const { basket } = message;
-          importBasket(basket, sendResponse);
+          importBasket(basket);
           break;
         default:
           sendResponse({ success: false, message: 'Invalid message type' });
@@ -134,7 +98,7 @@ const checkForElement = (selectorString: string, queryElement: HTMLElement = doc
       const itemLocatorElement = findAll
         ? queryElement.querySelectorAll(selectorString)
         : queryElement.querySelector(selectorString);
-      if (itemLocatorElement || (findAll && itemLocatorElement.length > 0)) {
+      if (itemLocatorElement || (findAll && (itemLocatorElement || []).length > 0)) {
         clearInterval(itemLocatorIntervalId);
         count = 0;
         resolve(itemLocatorElement);
@@ -148,21 +112,16 @@ const checkForElement = (selectorString: string, queryElement: HTMLElement = doc
   });
 };
 
-const invItemSelect = async (region: string, item: string, isAlsoSell = false, ChildItemID = 0) => {
-  //TKS 12.09.2020rgb(95, 102, 91) I have to change the name/id of this field because the 'Add new part' link that pulls up
-  //the add item control from inventory also has a hid_itemid and breaks because we have 2 idental hidden fields
+const invItemSelect = async (region: string, item: string, isAlsoSell: boolean = false, ChildItemID: number = 0) => {
   (document.querySelector('#hid_invoice_itemid') as HTMLInputElement).value = item;
 
-  //TKS 05.15.2015 #66724 added this hidden field to track child
   (document.querySelector('#hid_childitemid') as HTMLInputElement).value = `${ChildItemID}`;
 
-  //TKS 06.18.2014 #53411
   await closeInvoiceItemlocate(region);
   (document.querySelector('#txt_descriptiontoadd') as HTMLTextAreaElement).focus();
 
-  //TKS 02.21.2012 #12364 added lost sale button. Hide by default until they select an item
   (document.querySelector('#hideshow_lostsale') as HTMLDivElement).style.display = 'inline';
-  return await getNewInvoicePartInfo(isAlsoSell, ChildItemID);
+  return await getNewInvoicePartInfo(ChildItemID);
 };
 
 const closeInvoiceItemlocate = async (elementId: string) => {
@@ -175,28 +134,25 @@ const closeInvoiceItemlocate = async (elementId: string) => {
   return '';
 };
 
-const getNewInvoicePartInfo = async (isAlsoSell = false, ChildItemID = 0) => {
+const getNewInvoicePartInfo = async (ChildItemID: number = 0) => {
   const globalInvoiceID = 0;
   //TKS 12.09.2020 #183305 I have to change the name/id of this field because the 'Add new part' link that pulls up
   //the add item control from inventory also has a hid_itemid and breaks because we have 2 idental hidden fields
   let itemId = 0;
-  const hiddenInvoiceItemId = (document.querySelector('#hid_invoice_itemid') as HTMLInputElement).value || 0;
+  const hiddenInvoiceItemId = parseInt(
+    (document.querySelector('#hid_invoice_itemid') as HTMLInputElement).value || '0',
+  );
   if (hiddenInvoiceItemId > 0) {
     itemId = hiddenInvoiceItemId;
   }
 
   const partNumber = encodeURIComponent((document.querySelector('#txt_partnumbertoadd') as HTMLInputElement)?.value);
   const contactId = (document.querySelector('#global_contactid') as HTMLInputElement)?.value;
-  // AP: TODO: Unsure if this ever will need to change
 
-  //11.09.2011 jss - added encodeURIComponent cuz salina has pound signs in their part #'s
   const path = `invoicing/newinvoice/cs_invoice_items.php?partnumber=${partNumber}&contactid=${contactId}&itemid=${itemId}
     &invtype=${invoice_type}&ChildItemID=${ChildItemID}&invoiceid=${globalInvoiceID}`;
 
-  return await getData(path).then(response => {
-    return response;
-    console.log('response', response);
-  });
+  return await getData(path);
 };
 
 const call_cs_invoice_items = async (response: string) => {
@@ -220,8 +176,6 @@ const call_cs_invoice_items = async (response: string) => {
   const discountToAddInput = document.querySelector('#txt_discounttoadd') as HTMLInputElement;
 
   if (response.substring(0, 5) != 'blank') {
-    //07.05.2012 ghh - added to deal with json array instead of pipe array
-    //due to pipe symbols in descriptions
     myArray = JSON.parse(response);
 
     const hiddenPriceInput = document.querySelector('#hid_pricetoadd') as HTMLInputElement;
@@ -229,14 +183,11 @@ const call_cs_invoice_items = async (response: string) => {
     const toBePulledDiv = document.querySelector('#reg_ToBePulled') as HTMLDivElement;
     const hiddenMatrixLevelInput = document.querySelector('#hid_matrixlevel') as HTMLInputElement;
     const textOnOrderTextArea = document.querySelector('#txt_onorder') as HTMLTextAreaElement;
-    //07.05.2012 ghh - replaced above code with new json array parsing
-    descriptionTextArea.value = myArray[0];
 
+    descriptionTextArea.value = myArray[0];
     priceInput.value = myArray[2];
     hiddenPriceInput.value = myArray[2];
 
-    //TKS 12.09.2020 #183305 I have to change the name/id of this field because the 'Add new part' link that pulls up
-    //the add item control from inventory also has a hid_itemid and breaks because we have 2 idental hidden fields
     hiddenInvoiceItemInput.value = myArray[3];
     partNoteDiv.innerHTML = myArray[6];
     hiddenCostInput.value = myArray[7];
@@ -245,15 +196,15 @@ const call_cs_invoice_items = async (response: string) => {
     toBePulledDiv.innerHTML = myArray[10];
     hiddenMatrixLevelInput.value = myArray[11];
     textOnOrderTextArea.value = myArray[14];
+
     (document.querySelector('#hid_isRental') as HTMLInputElement).value = myArray[15];
-    //TKS 11.30.2022 #228241
+
     try {
       (document.querySelector('#hid_qoh') as HTMLInputElement).value = myArray[16];
     } catch (err) {}
     try {
       (document.querySelector('#hid_minorderqty') as HTMLInputElement).value = myArray[17];
     } catch (err) {}
-    //***********************
 
     try {
       discountToAddInput.value = myArray[9];
@@ -262,10 +213,7 @@ const call_cs_invoice_items = async (response: string) => {
     const poFieldDiv = document.querySelector('#po_field') as HTMLDivElement;
     poFieldDiv.style.display = myArray[4] == 1 ? 'inline' : 'none';
 
-    //TKS 06.04.2015 #68029 we now send back from the cs_invoice_items whether this field
-    //is readonly or not. It is not only contingent on security but if giftcard
     priceInput.readOnly = myArray[13] == 1 ? true : false;
-
     // close_invoice_itemlocate();
   } else {
     partNumberToAddInput.value = '';
@@ -275,19 +223,16 @@ const call_cs_invoice_items = async (response: string) => {
     hiddenCostInput.value = '';
     partNoteDiv.innerHTML = '';
     packagingTypeSpan.innerHTML = '';
-    //TKS 02.05.2015 #63040 added these regions to clear if blank comes back
-    //******************
+
     (document.querySelector('#reg_itemtaxes') as HTMLDivElement).innerHTML = '';
     (document.querySelector('#reg_driver_po') as HTMLDivElement).innerHTML = '';
     (document.querySelector('#reg_reference') as HTMLDivElement).innerHTML = '';
     (document.querySelector('#reg_replaceby') as HTMLDivElement).innerHTML = '';
     (document.querySelector('#reg_alsosell') as HTMLDivElement).innerHTML = '';
     (document.querySelector('#reg_bins') as HTMLSpanElement).innerHTML = '';
-    //TKS 07.11.2014 #54540 added sales promo region for potential promotions
     (document.querySelector('#reg_part_promos') as HTMLDivElement).innerHTML = '';
   }
 
-  //07.05.2012 ghh - changed broken to myarray
   if (myArray) {
     return await get_cv_invoice_itembins('reg_bins', myArray[3], '', '');
   } else {
@@ -296,12 +241,9 @@ const call_cs_invoice_items = async (response: string) => {
 };
 
 async function get_cv_invoice_itembins(placement: string, itemid: string, alllocations: string, otherstores: string) {
-  //TKS 9-15-10 #1636 this var gets passed in as 'all' if they click on the link to view other locations
   alllocations ??= '';
-  //10.12.2011 jss - if allx stores works like alllocations except it deals with multistore
   otherstores ??= '';
 
-  //11.27.2017 tam - 113263 adding jobid to this call
   const dropJobsSelect = document.querySelector('#drp_jobs') as HTMLSelectElement;
   const jobid = dropJobsSelect?.value || 0;
 
@@ -311,7 +253,7 @@ async function get_cv_invoice_itembins(placement: string, itemid: string, allloc
   const path = `invoicing/newinvoice/cv_invoice_itembins.php?placement=${placement}&alllocations=${alllocations}&itemid=${itemid}&invoiceid=${globalInvoiceID}
     &jobid=${jobid}&contactid=${globalContactId}&invoice_type=${invoice_type}&otherstores=${otherstores}`;
 
-  return await getData(path).then(response => {
+  return await getData(path).then((response: string) => {
     if (response.substring(0, 5) != 'Error') {
       const regBinsElement = document.querySelector('#reg_bins') as HTMLDivElement;
       regBinsElement.innerHTML = response;
@@ -332,12 +274,17 @@ const getData = async (url: string) => {
 
     return await response.text();
     // TODO: only want 1 request at a time, their logic tries to prevent this with globalGetItemList
-  } catch (error: unknown) {
+  } catch (error) {
     // TODO: extension probably has api to display error message better
     // TODO: add logic to save error to cache so I can review later
-    console.error(JSON.stringify(error.message));
-    return '';
+    if (error instanceof Error) {
+      console.error(JSON.stringify(error.message));
+    } else {
+      console.error(JSON.stringify(error));
+    }
   }
+
+  return '';
 };
 
 const findNonMiscQuantityElement = async () => {
