+++ pages/content/src/index.ts
@@ -1,8 +1,6 @@
 import { exportBasket } from '@src/exporter';
 import { addMessageListener } from '@src/importer';
 
-console.log('content script loaded - exporter');
-
 const currentUrl = window.location.href;
 const lizzyRegex = /[a-zA-Z0-9]*?\.nizex\.com/;
