--- pages/content/src/sampleFunction.ts
@@ -1,3 +1,0 @@
-export function sampleFunction() {
-  console.log('content script - sampleFunction() called from another module');
-}
