+++ pages/content/tsconfig.json
@@ -7,5 +7,5 @@
     },
     "types": ["chrome", "node"]
   },
-  "include": ["src", "vite.config.mts"]
+  "include": ["src", "vite.config.mts", "../../chrome-extension/public/types.ts"]
 }
